refactor on preferences http get call
